[REF]Move files without lmerging into models * Move file * FIxe import in differents modules
