+++ prestashoperpconnect/__init__.py
@@ -22,13 +22,13 @@
 #
 ###############################################################################
 
-import backend
+#import backend
 import connector
-import product
+#import product
 import product_combination
 import sale
 import delivery
-import consumer
+#import consumer
 import account
 import models
-import stock_tracking
+#import stock_tracking
