+++ prestashoperpconnect/models/stock_tracking.py
@@ -28,9 +28,9 @@ from openerp.addons.connector.queue.job import job
 from openerp.addons.connector.exception import FailedJobError, NoExternalId
 from openerp.addons.connector.unit.synchronizer import ExportSynchronizer
 from openerp.addons.connector_ecommerce.event import on_tracking_number_added
-from .connector import get_environment
+from ..connector import get_environment
 from .backend import prestashop
-from .unit.backend_adapter import PrestaShopCRUDAdapter
+from ..unit.backend_adapter import PrestaShopCRUDAdapter
 
 _logger = logging.getLogger(__name__)
