+++ prestashoperpconnect_catalog_manager/models/product.py
@@ -42,9 +42,9 @@ from openerp.addons.prestashoperpconnect.unit.mapper import \
     PrestashopExportMapper
 
 from openerp.addons.prestashoperpconnect.connector import get_environment
-from openerp.addons.prestashoperpconnect.backend import prestashop
-from openerp.addons.prestashoperpconnect.product import INVENTORY_FIELDS
-from openerp.addons.prestashoperpconnect_catalog_manager.product_combination \
+from openerp.addons.prestashoperpconnect.models.backend import prestashop
+from openerp.addons.prestashoperpconnect.models.product_backend import INVENTORY_FIELDS
+from .product_combination \
     import product_product_write
 from openerp.osv import fields, orm
 import openerp.addons.decimal_precision as dp
