+++ prestashoperpconnect_customize_example/models/backend.py
@@ -22,7 +22,7 @@
 
 
 import openerp.addons.connector.backend as backend
-import openerp.addons.prestashoperpconnect.backend as prestashop_backend
+import openerp.addons.prestashoperpconnect.models.backend as prestashop_backend
 
 prestashop_myversion = backend.Backend(parent=prestashop_backend.prestashop1500,
                                     version='1.5-myversion')
